adjusting crud

+++ WKDomain/ModelViews/Produto/NovoProduto.cs
@@ -30,7 +30,7 @@ namespace WKDomain.ModelViews
         /// <example>1/example>
         [Required(ErrorMessage = "{0} obrigatória")]
         [Display(Name = "Categoria")]
-        public decimal CategoriaId { get; set; }
+        public int CategoriaId { get; set; }
 
         public Categoria Categoria { get; set; }
